// src/api/model/editAddressDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 互帮Api
 * OpenAPI spec version: v1
 */

export interface EditAddressDto {
  /**
   * 街道/小区
   * @nullable
   */
  street?: string | null;
  /**
   * 楼号/门牌号
   * @nullable
   */
  unit?: string | null;
  /**
   * 联系人
   * @nullable
   */
  contact?: string | null;
  /**
   * 手机
   * @nullable
   */
  phone?: string | null;
  id?: number;
}